refactor: Fix circular dependency issue.
