+++ integration-libs/cdc/src/auth/facade/cdc-auth.service.ts
@@ -9,11 +9,11 @@ import {
   GlobalMessageService,
   GlobalMessageType,
   OCC_USER_ID_CURRENT,
-  UserIdService,
+  UserIdService
 } from '@spartacus/core';
 import { combineLatest, of } from 'rxjs';
 import { map, take, tap } from 'rxjs/operators';
-import { CdcAuthActions } from '../store/actions';
+import { CdcAuthActions } from '../store/actions/index';
 
 /**
  * Service to support custom CDC OAuth flow.
