+++ integration-libs/cdc/src/auth/store/cdc-auth-store.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { EffectsModule } from '@ngrx/effects';
-import { CdcUserTokenEffects } from './effects';
+import { CdcUserTokenEffects } from './effects/cdc-user-token.effect';
 
 @NgModule({
   imports: [CommonModule, EffectsModule.forFeature([CdcUserTokenEffects])],
