+++ integration-libs/cdc/src/auth/store/effects/cdc-user-token.effect.ts
@@ -7,14 +7,16 @@ import {
 } from '@spartacus/core';
 import { EMPTY, Observable, of } from 'rxjs';
 import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
-import { CdcAuthService } from '../../facade';
+import { CdcAuthService } from '../../facade/cdc-auth.service';
 import { CdcUserAuthenticationTokenService } from '../../services/user-authentication/cdc-user-authentication-token.service';
-import { CdcAuthActions } from '../actions';
+import { CdcAuthActions } from '../actions/index';
 
 @Injectable()
 export class CdcUserTokenEffects {
   @Effect()
-  loadCdcUserToken$: Observable<CdcAuthActions.CdcUserTokenAction> = this.actions$.pipe(
+  loadCdcUserToken$: Observable<
+    CdcAuthActions.CdcUserTokenAction
+  > = this.actions$.pipe(
     ofType(CdcAuthActions.LOAD_CDC_USER_TOKEN),
     map((action: CdcAuthActions.LoadCdcUserToken) => action.payload),
     mergeMap((payload) =>
